# Use mt_mode 1 for improved blast performance in transporter.sh